feat: add keyword filter link to project page
