+++ frontend/components/software/SoftwareKeywords.tsx
@@ -6,6 +6,7 @@
 import LocalOfferIcon from '@mui/icons-material/LocalOffer'
 import {KeywordForSoftware} from '../../types/SoftwareTypes'
 import TagChipFilter from '../layout/TagChipFilter'
+import {softwareUrl} from '~/utils/postgrestUrl'
 
 export default function SoftwareKeywords({keywords = []}: { keywords: KeywordForSoftware[] }) {
 
@@ -18,7 +19,8 @@ export default function SoftwareKeywords({keywords = []}: { keywords: KeywordFor
     return (
       <div className="flex flex-wrap py-1">
         {keywords.map((item, pos) => {
-          return <TagChipFilter key={pos} label={item.keyword} />
+          const url = softwareUrl({keywords: [item.keyword]})
+          return <TagChipFilter url={url} key={pos} label={item.keyword} />
         })}
       </div>
     )
